excepciones en busqueda de pedido get/delete implementadas

+++ ms-pedidos/src/main/java/isi/dan/ms/pedidos/controller/PedidoController.java
@@ -1,11 +1,17 @@
 package isi.dan.ms.pedidos.controller;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import isi.dan.ms.pedidos.Exceptions.ClienteNotFoundException;
+import isi.dan.ms.pedidos.Exceptions.PedidoNotFoundException;
 import isi.dan.ms.pedidos.modelo.EstadoPedido;
 import isi.dan.ms.pedidos.modelo.Pedido;
 import isi.dan.ms.pedidos.servicio.PedidoService;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.List;
 
@@ -16,6 +22,9 @@ public class PedidoController {
     
     @Autowired
     private PedidoService pedidoService;
+
+    Logger log = LoggerFactory.getLogger(PedidoController.class);
+
 
     @PostMapping
     public ResponseEntity<Pedido> createPedido(@RequestBody Pedido pedido) {
@@ -29,15 +38,25 @@ public class PedidoController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Pedido> getPedidoById(@PathVariable String id) {
+    public ResponseEntity<Pedido> getPedidoById(@PathVariable String id) throws PedidoNotFoundException {
         Pedido pedido = pedidoService.getPedidoById(id);
-        return pedido != null ? ResponseEntity.ok(pedido) : ResponseEntity.notFound().build();
+        if(pedido !=null){
+           return ResponseEntity.ok(pedido);
+        }else{
+            throw new PedidoNotFoundException("Pedido '" + id + "' no encontrado");
+        }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePedido(@PathVariable String id) {
-        pedidoService.deletePedido(id);
+    public ResponseEntity<Void> deletePedido(@PathVariable String id) throws PedidoNotFoundException {
+        Pedido pedido = pedidoService.getPedidoById(id);
+        if(pedido !=null){
+            pedidoService.deletePedido(id);
         return ResponseEntity.noContent().build();
+         }else{
+             throw new PedidoNotFoundException("Pedido '" + id + "' no encontrado");
+         }
+        
     }
 }
